king: implement -N, dry-run mode.
When -N is enabled, no sockets are bound and no events get persisted
to the datastore. We also pass the dry run flag to the serf, who
should not snapshot.

(For redundancy, we should probably also make king not send the save
snapshot commands, but I tested locally that the worker process
doesn't save.)