king: add partial-replay and --dry-from
partial-replay allows you to replay up to event N and make a checkpoint
there.  --dry-from replays to event N and then starts in dry-run mode
from that point, so you can attach a console.

The intended debug flow is to use partial-replay to make a checkpoint
near when a bug is triggered, then use event-browser and --dry-from to
drill down and find exactly what triggered the bug.

Unfortunately, if the bug was triggered by crash handling, you do not
have the original event, only the injected replacement event.  This is
still useful, though.